Fix weird error in Windows' Arduino IDE
It did not like variable names error nor Serial1 as they seemed to be already taken for other uses. It only caused trouble in Windows and OSX and Mac versions of Arduino IDE did not complain. 

This new way should please every version of the IDE and it is tested to work in Windows 10.